+++ changelog.md
@@ -1,3 +1,6 @@
+latest
+ - Store list in indexedDb for optional restoring of the last session on next load.
+
 2023.10.15 1.4.8
  - New serialize selected action, any selected samples that are stereo files, will have their stereo channels serialized to mono, by Left/Right, Side/Mid, or Left/Right/Side/Mid.
  - Dual mono exports setting error.